segmentation/requirements.txt
# Segmentation Project Requirements
# Install after root requirements: pip install -r requirements.txt

# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0

# Image processing and visualization
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Segmentation specific
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0

# Experiment tracking and monitoring
wandb>=0.15.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
omegaconf>=2.3.0

# Dataset handling
datasets>=2.14.0

# Metrics and evaluation
scikit-learn>=1.3.0
torchmetrics>=1.0.0
